+++ eth_defi/erc_4626/core.py
@@ -239,6 +239,11 @@ class ERC4626Feature(enum.Enum):
     #: https://etherscan.io/address/0x3ed6aa32c930253fc990de58ff882b9186cd0072
     cap_like = "cap_like"
 
+    #: Foxify
+    #:
+    #: https://www.foxify.trade/
+    foxify_like = "foxify_like"
+
 
 def get_vault_protocol_name(features: set[ERC4626Feature]) -> str:
     """Deduct vault protocol name based on Vault smart contract features.
@@ -345,6 +350,9 @@ def get_vault_protocol_name(features: set[ERC4626Feature]) -> str:
 
     elif ERC4626Feature.cap_like in features:
         return "CAP"
+
+    elif ERC4626Feature.foxify_like in features:
+        return "Foxify"
 
     # No idea
     if ERC4626Feature.erc_7540_like in features: