+++ eth_defi/foxify/__init__.py
@@ -1,0 +1,1 @@
+"""Foxify protocol integration."""